+++ src/features/goblins/village/Village.tsx
@@ -8,6 +8,7 @@ import { Bank } from "../bank/Bank";
 import { Farm } from "../farm/Farm";
 import { WishingWell } from "../wishingWell/WishingWell";
 import { Hud } from "../hud/Hud";
+import { Market } from "../market/Market";
 
 export const Village: React.FC = () => {
   return (
@@ -30,6 +31,7 @@ export const Village: React.FC = () => {
       <Bank />
       <Farm />
       <WishingWell />
+      <Market />
     </div>
   );
 };
